chore: Rename missionsPrototype page folder to missions
The folder name called the Storybook work a prototype. Use missions
so the page name matches the product term.
