+++ web_src/src/pages/factories/pages/missions/MissionDetailHeader.tsx
