+++ web_src/src/pages/factories/pages/missions/MissionDetailPage.tsx
