+++ web_src/src/pages/factories/pages/missions/MissionStatusValue.tsx
