+++ web_src/src/pages/factories/pages/missions/WorkOrderMissionOverviewRow.tsx
