+++ web_src/src/pages/factories/pages/missions/missionPaths.ts
